+++ examples/src/config.ml
@@ -1,0 +1,4 @@
+let (@+) x y = x + y
+
+type t = int
+let t = 2