+++ examples/src/config.mli
@@ -1,0 +1,4 @@
+val (@+) : int -> int -> int
+
+type t
+val t : t